requirements.txt
Django==1.7
django-crispy-forms
solid_i18n
pytz
unidecode
django-vinaigrette
django-grappelli
django-import-export
pillow
bcrypt
